backend/pyproject.toml
[project]
name = "spectertest-backend"
version = "0.1.0"
description = "SpecterTest — Swarm Intelligence Behavioral Testing Platform"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=5.0.0",
    "flask-sqlalchemy>=3.0.5",
    "playwright>=1.51.0",
    "openai>=1.60.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
